Move cache rebuilding to the codegen module
